requirements.txt
celery==4.2.1
Django==2.1.7
djangorestframework==3.9.1
mock==2.0.0
redis==3.2.0
